Changes in the encoder and preview window - the encoder has been moved to a separate class, - the preview image is resized when the preview window is resized.
