+++ app/mainwindow.h
@@ -44,11 +44,6 @@
 #include <QHBoxLayout>
 #include <QGridLayout>
 #include <QLabel>
-#include <iostream>
-#include <iomanip>
-#include <sstream>
-#include <ctime>
-#include <math.h>
 #include <QMovie>
 #include <QDockWidget>
 #include <QFile>
@@ -57,7 +52,14 @@
 #include <QSettings>
 #include <QMap>
 #include <QTranslator>
+#include <iostream>
+#include <iomanip>
+#include <sstream>
+#include <ctime>
+#include <math.h>
 #include "openingfiles.h"
+#include "encoder.h"
+#include "constants.h"
 
 
 
@@ -79,9 +81,7 @@ public:
 
     ~Widget();
 
-    static const int PARAMETERS_COUNT = 34;
-
-    // ******************** Settings *************************//
+    // ============= Settings ================
 
     bool        _hideInTrayFlag,
                 _showHDR_mode,
@@ -160,25 +160,31 @@ private slots:
 
     QString callFileDialog(const QString title);
 
-    void make_preset();
+    // ============= Encoder ================
+
+    void initEncoding();
+
+    void onEncodingMode(const QString &mode);
+
+    void onEncodingStarted();
+
+    void onEncodingInitError(const QString &_message);
+
+    void onEncodingProgress(const int &percent, const float &rem_time);
+
+    void onEncodingLog(const QString &log);
+
+    void onEncodingCompleted();
+
+    void onEncodingAborted();
+
+    void onEncodingError(const QString &error_message);
 
     void pause();
 
     void resume();
 
-    void add_metadata();
-
-    void complete();
-
-    void progress_1();
-
-    void progress_2();
-
-    void cancel();
-
-    void error();
-
-    void encode();
+    // ============= Video Metadata ================
 
     void on_lineEditTitleVideo_editingFinished();
 
@@ -266,29 +272,17 @@ private:
 
     Ui::Widget *ui;
 
+    Encoder *encoder;
+
     QVector<QString> _cur_param;
 
     QVector<QVector<QString>> _preset_table;
 
     OpeningFiles openingFiles;
 
-    // ******************* Constants *****************************//
+    QPixmap preview_pixmap;
 
-    static const int PRESETS_VERSION = 352;
-
-    static const int SETTINGS_VERSION = 352;
-
-    static const int NUMBER_PRESETS = 29;
-
-    static const int AMOUNT_AUDIO_STREAMS = 9;
-
-    static const int AMOUNT_SUBTITLES = 9;
-
-    static constexpr float MAXIMUM_ALLOWED_TIME = 359999.0f;
-
-    // ******************** Dock area ****************************//
-
-    static const int DOCKS_COUNT = 8;
+    // ============= Dock area =============
 
     QMainWindow *window;
 
@@ -296,17 +290,17 @@ private:
 
     QDockWidget *docks[DOCKS_COUNT];
 
-    // ******************** Top label ****************************//
+    // ============= Top label =============
 
     QLabel      *raiseThumb,
                 *audioThumb,
                 *subtitleThumb;
 
-    // **************** Progress animation ***********************//
+    // ============= Progress animation =============
 
     QMovie      *animation;
 
-    // ***************** Tray menu actions ************************//
+    // ============= Tray menu actions =============
 
     QSystemTrayIcon *trayIcon;
 
@@ -316,7 +310,7 @@ private:
                 *restoreAction,
                 *quitAction;
 
-    // ***************** Top menu actions ************************//
+    // ============= Top menu actions =============
 
     QAction     *add_files,
                 *remove_files,
@@ -343,33 +337,32 @@ private:
                 *menuPreferences,
                 *menuAbout;
 
-    // ***************** Table menu actions ************************//
+    // ============= Table menu actions =============
 
     QMenu       *itemMenu;
 
-    // ****************** Tree menu actions ************************//
+    // ============= Tree menu actions =============
 
     QMenu       *sectionMenu,
                 *presetMenu;
 
-    // ***************** Preset menu actions ***********************//
+    // ============= Preset menu actions =============
 
     QAction     *addsection,
                 *addpreset;
 
     QMenu       *menu;
 
-    // ********************** Processes ****************************//
+    // ============= Processes =============
 
-    QProcess    *processEncoding,
-                *processThumbCreation;
+    QProcess    *processThumbCreation;
 
-    // *********************** Timers ******************************//
+    // ============= Timers =============
 
     QTimer      *timer,
                 *timerCallSetThumbnail;
 
-    // ******************** Initialization *************************//
+    // ============= Initialization =============
 
     QFile       _wind_file;
 
@@ -377,7 +370,7 @@ private:
 
     int         _theme;
 
-    // ************************** Metadata ************************//
+    // ============= Metadata =============
 
     int         _audioStreamCheckState[AMOUNT_AUDIO_STREAMS],
                 _subtitleCheckState[AMOUNT_SUBTITLES];
@@ -389,19 +382,15 @@ private:
                 _subtitleLang[AMOUNT_SUBTITLES],
                 _subtitleTitle[AMOUNT_SUBTITLES];
 
-    // ************************* Encoding ***************************//
+    // ============= Encoding =============
 
-    bool        _flag_two_pass,
-                _flag_hdr,
-                _calling_pr_1,
-                _mux_mode,
-                _batch_mode;
+    bool        _batch_mode;
 
     int         _fr_count,
+                _status_encode_btn,
                 _row;
 
-    time_t      _loop_start,
-                _strt_t;
+    time_t      _strt_t;
 
     double      _dur,
                 _curTime,
@@ -409,7 +398,6 @@ private:
                 _endTime;
 
     QString     _message,
-                _error_message,
                 _curFilename,
                 _curPath,
                 _temp_file,
@@ -418,26 +406,16 @@ private:
                 _settings_path,
                 _thumb_path,
                 _preset_file,
-                _preset_0,
-                _preset_pass1,
-                _preset,
-                _preset_mkvmerge,
-                _sub_mux_param,
                 _fmt,
                 _width,
                 _height,
                 _fps,
                 _stream_size,
-                _status_encode_btn;
+                extension;
 
     QSettings   *_settings;
 
-    // *********************** Geometry ***********************************//
-
-    enum Resize {
-        LEFT,   LEFT_TOP,   LEFT_BOTTOM,   TOP,
-        RIGHT,  RIGHT_TOP,  RIGHT_BOTTOM,  BOTTOM
-    };
+    // ============= Geometry =============
 
     bool        _windowActivated,
                 _expandWindowsState,
@@ -454,7 +432,7 @@ private:
 
     QVector<bool> _clickPressedToResizeFlag;    
 
-    // ****************************************************************************//
+    // ====================================
 
     void createConnections();
 
@@ -488,9 +466,12 @@ private:
 
     QString timeConverter(double &time);
 
-    QString timeConverter(float &time);
+    QString timeConverter(const float &time);
 
-    QString setThumbnail(QString curFilename, double time, QString quality, int destination);
+    QString setThumbnail(QString curFilename,
+                         const double &time,
+                         const int &quality,
+                         const int &destination);
 
     void provideContextMenu(const QPoint &position);
 
@@ -500,7 +481,7 @@ private:
 
     void resetView();
 
-    /************************* Preset Window ******************************/
+    // ============= Preset Window =============
 
     void set_defaults();
 
@@ -508,9 +489,15 @@ private:
 
     void updateCurPresetPos(int &index_top, int &index_child);
 
-    void updateInfoFields(QString &codec_qstr, QString &mode_qstr, QString &container_qstr,
-                          QString &bqr_qstr, QString &pass_qstr, QString &preset_qstr,
-                          QString &acodec_qstr, QTreeWidgetItem *item, bool defaultNameFlag);
+    void updateInfoFields(QString &codec_qstr,
+                          QString &mode_qstr,
+                          QString &container_qstr,
+                          QString &bqr_qstr,
+                          QString &pass_qstr,
+                          QString &preset_qstr,
+                          QString &acodec_qstr,
+                          QTreeWidgetItem *item,
+                          bool defaultNameFlag);
 
     void updatePresetTable();
 
@@ -526,66 +513,6 @@ private:
 
     QString updateFieldContainer(int &codec, int &container);
 
-    enum ColumnIndex {
-        FILENAME,       FORMAT,         RESOLUTION,     DURATION,           FPS,
-        AR,             STATUS,         BITRATE,        SUBSAMPLING,        BITDEPTH,
-        COLORSPACE,     COLORRANGE,     COLORPRIM,      COLORMATRIX,        TRANSFER,
-        MAXLUM,         MINLUM,         MAXCLL,         MAXFALL,            MASTERDISPLAY,
-        PATH,           T_DUR,          T_CHROMACOORD,  T_WHITECOORD,       T_STREAMSIZE,
-        T_WIDTH,        T_HEIGHT,       T_VIDEOTITLE,   T_VIDEOMOVIENAME,   T_VIDEOYEAR,
-        T_VIDEOAUTHOR,  T_VIDEOPERF,    T_VIDEODESCR,   T_AUDIO_1,          T_AUDIO_2,
-        T_AUDIO_3,      T_AUDIO_4,      T_AUDIO_5,      T_AUDIO_6,          T_AUDIO_7,
-        T_AUDIO_8,      T_AUDIO_9,      T_AUDIOLANG_1,  T_AUDIOLANG_2,      T_AUDIOLANG_3,
-        T_AUDIOLANG_4,  T_AUDIOLANG_5,  T_AUDIOLANG_6,  T_AUDIOLANG_7,      T_AUDIOLANG_8,
-        T_AUDIOLANG_9,  T_AUDIOTITLE_1, T_AUDIOTITLE_2, T_AUDIOTITLE_3,     T_AUDIOTITLE_4,
-        T_AUDIOTITLE_5, T_AUDIOTITLE_6, T_AUDIOTITLE_7, T_AUDIOTITLE_8,     T_AUDIOTITLE_9,
-        T_AUDIOCHECK_1, T_AUDIOCHECK_2, T_AUDIOCHECK_3, T_AUDIOCHECK_4,     T_AUDIOCHECK_5,
-        T_AUDIOCHECK_6, T_AUDIOCHECK_7, T_AUDIOCHECK_8, T_AUDIOCHECK_9,     T_SUBTITLE_1,
-        T_SUBTITLE_2,   T_SUBTITLE_3,   T_SUBTITLE_4,   T_SUBTITLE_5,       T_SUBTITLE_6,
-        T_SUBTITLE_7,   T_SUBTITLE_8,   T_SUBTITLE_9,   T_SUBLANG_1,        T_SUBLANG_2,
-        T_SUBLANG_3,    T_SUBLANG_4,    T_SUBLANG_5,    T_SUBLANG_6,        T_SUBLANG_7,
-        T_SUBLANG_8,    T_SUBLANG_9,    T_TITLESUB_1,   T_TITLESUB_2,       T_TITLESUB_3,
-        T_TITLESUB_4,   T_TITLESUB_5,   T_TITLESUB_6,   T_TITLESUB_7,       T_TITLESUB_8,
-        T_TITLESUB_9,   T_SUBCHECK_1,   T_SUBCHECK_2,   T_SUBCHECK_3,       T_SUBCHECK_4,
-        T_SUBCHECK_5,   T_SUBCHECK_6,   T_SUBCHECK_7,   T_SUBCHECK_8,       T_SUBCHECK_9,
-        T_STARTTIME,    T_ENDTIME
-    };
-
-    enum CurParamIndex {
-        OUTPUT_PARAM,   CODEC,          MODE,
-        CONTAINER,      BQR,            MAXRATE,
-        BUFSIZE,        FRAME_RATE,     BLENDING,
-        WIDTH,          HEIGHT,         PASS,
-        PRESET,         COLOR_RANGE,    MIN_LUM,
-        MAX_LUM,        MAX_CLL,        MAX_FALL,
-        MASTER_DISPLAY, CHROMA_COORD,   WHITE_COORD,
-        AUDIO_CODEC,    AUDIO_BITRATE,  MINRATE,
-        LEVEL,          ASAMPLE_RATE,   ACHANNELS,
-        MATRIX,         PRIMARY,        TRC,
-        PRESET_NAME,    REP_PRIM,       REP_MATRIX,
-        REP_TRC
-    };
-
-    enum CurHDRIndex {
-        CUR_COLOR_RANGE,    CUR_COLOR_PRIMARY,  CUR_COLOR_MATRIX,
-        CUR_TRANSFER,       CUR_MAX_LUM,        CUR_MIN_LUM,
-        CUR_MAX_CLL,        CUR_MAX_FALL,       CUR_MASTER_DISPLAY,
-        CUR_CHROMA_COORD,   CUR_WHITE_COORD
-    };
-
-    enum VideoMetadataIndex {
-        VIDEO_TITLE,        VIDEO_AUTHOR,       VIDEO_YEAR,
-        VIDEO_PERFORMER,    VIDEO_DESCRIPTION,  VIDEO_MOVIENAME, ENUM_END_LABEL
-    };
-
-    enum DockIndex {
-        PRESETS_DOCK, PREVIEW_DOCK, SOURCE_DOCK, OUTPUT_DOCK,
-        STREAMS_DOCK, LOG_DOCK, METADATA_DOCK, SPLIT_DOCK
-    };
-
-    enum Theme {
-        GRAY, DARK, WAVE, DEFAULT
-    };
 };
 
 #endif // WIDGET_H
