+++ app/preset.cpp
@@ -344,39 +344,39 @@ void Preset::setParameters(QByteArray *ptr_presetWindowGeometry, QVector<QString
 //    ui->lineEdit_max_cll->setValidator(intValidator);
 //    ui->lineEdit_max_fall->setValidator(intValidator);
 
-    ui->textBrowser_presetname->setText((*_new_param)[Parameters::_OUTPUT_PARAM]);
-    ui->comboBox_codec->setCurrentIndex((*_new_param)[Parameters::_CODEC].toInt());
-    ui->comboBox_mode->setCurrentIndex((*_new_param)[Parameters::_MODE].toInt());
-    ui->comboBox_container->setCurrentIndex((*_new_param)[Parameters::_CONTAINER].toInt());
-    ui->lineEdit_bitrate->setText((*_new_param)[Parameters::_BQR]);
-    ui->lineEdit_minrate->setText((*_new_param)[Parameters::_MINRATE]);
-    ui->lineEdit_maxrate->setText((*_new_param)[Parameters::_MAXRATE]);
-    ui->lineEdit_bufsize->setText((*_new_param)[Parameters::_BUFSIZE]);
-    ui->comboBox_level->setCurrentIndex((*_new_param)[Parameters::_LEVEL].toInt());
-    ui->comboBoxFrameRate->setCurrentIndex((*_new_param)[Parameters::_FRAME_RATE].toInt());
-    ui->comboBoxBlending->setCurrentIndex((*_new_param)[Parameters::_BLENDING].toInt());
-    ui->comboBox_width->setCurrentIndex((*_new_param)[Parameters::_WIDTH].toInt());
-    ui->comboBox_height->setCurrentIndex((*_new_param)[Parameters::_HEIGHT].toInt());
-    ui->comboBox_pass->setCurrentIndex((*_new_param)[Parameters::_PASS].toInt());
-    ui->comboBox_preset->setCurrentIndex((*_new_param)[Parameters::_PRESET].toInt());
-    ui->comboBox_color_range->setCurrentIndex((*_new_param)[Parameters::_COLOR_RANGE].toInt());
-    ui->comboBox_color_prim->setCurrentIndex((*_new_param)[Parameters::_PRIMARY].toInt());
-    ui->comboBox_color_matrix->setCurrentIndex((*_new_param)[Parameters::_MATRIX].toInt());
-    ui->comboBox_transfer->setCurrentIndex((*_new_param)[Parameters::_TRC].toInt());
-    ui->lineEdit_min_lum->setText((*_new_param)[Parameters::_MIN_LUM]);
-    ui->lineEdit_max_lum->setText((*_new_param)[Parameters::_MAX_LUM]);
-    ui->lineEdit_max_cll->setText((*_new_param)[Parameters::_MAX_CLL]);
-    ui->lineEdit_max_fall->setText((*_new_param)[Parameters::_MAX_FALL]);
-    ui->comboBox_master_disp->setCurrentIndex((*_new_param)[Parameters::_MASTER_DISPLAY].toInt());
-    ui->lineEdit_chroma_coord->setText((*_new_param)[Parameters::_CHROMA_COORD]);
-    ui->lineEdit_white_coord->setText((*_new_param)[Parameters::_WHITE_COORD]);
-    ui->comboBox_audio_codec->setCurrentIndex((*_new_param)[Parameters::_AUDIO_CODEC].toInt());
-    ui->comboBox_audio_bitrate->setCurrentIndex((*_new_param)[Parameters::_AUDIO_BITRATE].toInt());
-    ui->comboBox_audio_sampling->setCurrentIndex((*_new_param)[Parameters::_ASAMPLE_RATE].toInt());
-    ui->comboBox_audio_channels->setCurrentIndex((*_new_param)[Parameters::_ACHANNELS].toInt());
-    ui->checkBox_primaries->setCheckState((Qt::CheckState)(*_new_param)[Parameters::_REP_PRIM].toInt());
-    ui->checkBox_matrix->setCheckState((Qt::CheckState)(*_new_param)[Parameters::_REP_MATRIX].toInt());
-    ui->checkBox_transfer->setCheckState((Qt::CheckState)(*_new_param)[Parameters::_REP_TRC].toInt());
+    ui->textBrowser_presetname->setText((*_new_param)[CurParamIndex::OUTPUT_PARAM]);
+    ui->comboBox_codec->setCurrentIndex((*_new_param)[CurParamIndex::CODEC].toInt());
+    ui->comboBox_mode->setCurrentIndex((*_new_param)[CurParamIndex::MODE].toInt());
+    ui->comboBox_container->setCurrentIndex((*_new_param)[CurParamIndex::CONTAINER].toInt());
+    ui->lineEdit_bitrate->setText((*_new_param)[CurParamIndex::BQR]);
+    ui->lineEdit_minrate->setText((*_new_param)[CurParamIndex::MINRATE]);
+    ui->lineEdit_maxrate->setText((*_new_param)[CurParamIndex::MAXRATE]);
+    ui->lineEdit_bufsize->setText((*_new_param)[CurParamIndex::BUFSIZE]);
+    ui->comboBox_level->setCurrentIndex((*_new_param)[CurParamIndex::LEVEL].toInt());
+    ui->comboBoxFrameRate->setCurrentIndex((*_new_param)[CurParamIndex::FRAME_RATE].toInt());
+    ui->comboBoxBlending->setCurrentIndex((*_new_param)[CurParamIndex::BLENDING].toInt());
+    ui->comboBox_width->setCurrentIndex((*_new_param)[CurParamIndex::WIDTH].toInt());
+    ui->comboBox_height->setCurrentIndex((*_new_param)[CurParamIndex::HEIGHT].toInt());
+    ui->comboBox_pass->setCurrentIndex((*_new_param)[CurParamIndex::PASS].toInt());
+    ui->comboBox_preset->setCurrentIndex((*_new_param)[CurParamIndex::PRESET].toInt());
+    ui->comboBox_color_range->setCurrentIndex((*_new_param)[CurParamIndex::COLOR_RANGE].toInt());
+    ui->comboBox_color_prim->setCurrentIndex((*_new_param)[CurParamIndex::PRIMARY].toInt());
+    ui->comboBox_color_matrix->setCurrentIndex((*_new_param)[CurParamIndex::MATRIX].toInt());
+    ui->comboBox_transfer->setCurrentIndex((*_new_param)[CurParamIndex::TRC].toInt());
+    ui->lineEdit_min_lum->setText((*_new_param)[CurParamIndex::MIN_LUM]);
+    ui->lineEdit_max_lum->setText((*_new_param)[CurParamIndex::MAX_LUM]);
+    ui->lineEdit_max_cll->setText((*_new_param)[CurParamIndex::MAX_CLL]);
+    ui->lineEdit_max_fall->setText((*_new_param)[CurParamIndex::MAX_FALL]);
+    ui->comboBox_master_disp->setCurrentIndex((*_new_param)[CurParamIndex::MASTER_DISPLAY].toInt());
+    ui->lineEdit_chroma_coord->setText((*_new_param)[CurParamIndex::CHROMA_COORD]);
+    ui->lineEdit_white_coord->setText((*_new_param)[CurParamIndex::WHITE_COORD]);
+    ui->comboBox_audio_codec->setCurrentIndex((*_new_param)[CurParamIndex::AUDIO_CODEC].toInt());
+    ui->comboBox_audio_bitrate->setCurrentIndex((*_new_param)[CurParamIndex::AUDIO_BITRATE].toInt());
+    ui->comboBox_audio_sampling->setCurrentIndex((*_new_param)[CurParamIndex::ASAMPLE_RATE].toInt());
+    ui->comboBox_audio_channels->setCurrentIndex((*_new_param)[CurParamIndex::ACHANNELS].toInt());
+    ui->checkBox_primaries->setCheckState((Qt::CheckState)(*_new_param)[CurParamIndex::REP_PRIM].toInt());
+    ui->checkBox_matrix->setCheckState((Qt::CheckState)(*_new_param)[CurParamIndex::REP_MATRIX].toInt());
+    ui->checkBox_transfer->setCheckState((Qt::CheckState)(*_new_param)[CurParamIndex::REP_TRC].toInt());
 }
 
 void Preset::on_closeWindow_clicked()
@@ -403,39 +403,39 @@ void Preset::on_buttonCancel_clicked()  /*** Close preset window ***/
 
 void Preset::on_buttonApply_clicked()  /*** Apply preset ***/
 {
-    (*_new_param)[Parameters::_OUTPUT_PARAM] = ui->textBrowser_presetname->toPlainText();
-    (*_new_param)[Parameters::_CODEC] = QString::number(ui->comboBox_codec->currentIndex());
-    (*_new_param)[Parameters::_MODE] = QString::number(ui->comboBox_mode->currentIndex());
-    (*_new_param)[Parameters::_CONTAINER] = QString::number(ui->comboBox_container->currentIndex());
-    (*_new_param)[Parameters::_BQR] = ui->lineEdit_bitrate->text();
-    (*_new_param)[Parameters::_MINRATE] = ui->lineEdit_minrate->text();
-    (*_new_param)[Parameters::_MAXRATE] = ui->lineEdit_maxrate->text();
-    (*_new_param)[Parameters::_BUFSIZE] = ui->lineEdit_bufsize->text();
-    (*_new_param)[Parameters::_LEVEL] = QString::number(ui->comboBox_level->currentIndex());
-    (*_new_param)[Parameters::_FRAME_RATE] = QString::number(ui->comboBoxFrameRate->currentIndex());
-    (*_new_param)[Parameters::_BLENDING] = QString::number(ui->comboBoxBlending->currentIndex());
-    (*_new_param)[Parameters::_WIDTH] = QString::number(ui->comboBox_width->currentIndex());
-    (*_new_param)[Parameters::_HEIGHT] = QString::number(ui->comboBox_height->currentIndex());
-    (*_new_param)[Parameters::_PASS]= QString::number(ui->comboBox_pass->currentIndex());
-    (*_new_param)[Parameters::_PRESET] = QString::number(ui->comboBox_preset->currentIndex());
-    (*_new_param)[Parameters::_COLOR_RANGE] = QString::number(ui->comboBox_color_range->currentIndex());
-    (*_new_param)[Parameters::_PRIMARY] = QString::number(ui->comboBox_color_prim->currentIndex());
-    (*_new_param)[Parameters::_MATRIX] = QString::number(ui->comboBox_color_matrix->currentIndex());
-    (*_new_param)[Parameters::_TRC] = QString::number(ui->comboBox_transfer->currentIndex());
-    (*_new_param)[Parameters::_MIN_LUM] = ui->lineEdit_min_lum->text();
-    (*_new_param)[Parameters::_MAX_LUM] = ui->lineEdit_max_lum->text();
-    (*_new_param)[Parameters::_MAX_CLL] = ui->lineEdit_max_cll->text();
-    (*_new_param)[Parameters::_MAX_FALL] = ui->lineEdit_max_fall->text();
-    (*_new_param)[Parameters::_MASTER_DISPLAY] = QString::number(ui->comboBox_master_disp->currentIndex());
-    (*_new_param)[Parameters::_CHROMA_COORD] = ui->lineEdit_chroma_coord->text();
-    (*_new_param)[Parameters::_WHITE_COORD] = ui->lineEdit_white_coord->text();
-    (*_new_param)[Parameters::_AUDIO_CODEC] = QString::number(ui->comboBox_audio_codec->currentIndex());
-    (*_new_param)[Parameters::_AUDIO_BITRATE] = QString::number(ui->comboBox_audio_bitrate->currentIndex());
-    (*_new_param)[Parameters::_ASAMPLE_RATE] = QString::number(ui->comboBox_audio_sampling->currentIndex());
-    (*_new_param)[Parameters::_ACHANNELS] = QString::number(ui->comboBox_audio_channels->currentIndex());
-    (*_new_param)[Parameters::_REP_PRIM] = QString::number(ui->checkBox_primaries->checkState());
-    (*_new_param)[Parameters::_REP_MATRIX] = QString::number(ui->checkBox_matrix->checkState());
-    (*_new_param)[Parameters::_REP_TRC] = QString::number(ui->checkBox_transfer->checkState());
+    (*_new_param)[CurParamIndex::OUTPUT_PARAM] = ui->textBrowser_presetname->toPlainText();
+    (*_new_param)[CurParamIndex::CODEC] = QString::number(ui->comboBox_codec->currentIndex());
+    (*_new_param)[CurParamIndex::MODE] = QString::number(ui->comboBox_mode->currentIndex());
+    (*_new_param)[CurParamIndex::CONTAINER] = QString::number(ui->comboBox_container->currentIndex());
+    (*_new_param)[CurParamIndex::BQR] = ui->lineEdit_bitrate->text();
+    (*_new_param)[CurParamIndex::MINRATE] = ui->lineEdit_minrate->text();
+    (*_new_param)[CurParamIndex::MAXRATE] = ui->lineEdit_maxrate->text();
+    (*_new_param)[CurParamIndex::BUFSIZE] = ui->lineEdit_bufsize->text();
+    (*_new_param)[CurParamIndex::LEVEL] = QString::number(ui->comboBox_level->currentIndex());
+    (*_new_param)[CurParamIndex::FRAME_RATE] = QString::number(ui->comboBoxFrameRate->currentIndex());
+    (*_new_param)[CurParamIndex::BLENDING] = QString::number(ui->comboBoxBlending->currentIndex());
+    (*_new_param)[CurParamIndex::WIDTH] = QString::number(ui->comboBox_width->currentIndex());
+    (*_new_param)[CurParamIndex::HEIGHT] = QString::number(ui->comboBox_height->currentIndex());
+    (*_new_param)[CurParamIndex::PASS]= QString::number(ui->comboBox_pass->currentIndex());
+    (*_new_param)[CurParamIndex::PRESET] = QString::number(ui->comboBox_preset->currentIndex());
+    (*_new_param)[CurParamIndex::COLOR_RANGE] = QString::number(ui->comboBox_color_range->currentIndex());
+    (*_new_param)[CurParamIndex::PRIMARY] = QString::number(ui->comboBox_color_prim->currentIndex());
+    (*_new_param)[CurParamIndex::MATRIX] = QString::number(ui->comboBox_color_matrix->currentIndex());
+    (*_new_param)[CurParamIndex::TRC] = QString::number(ui->comboBox_transfer->currentIndex());
+    (*_new_param)[CurParamIndex::MIN_LUM] = ui->lineEdit_min_lum->text();
+    (*_new_param)[CurParamIndex::MAX_LUM] = ui->lineEdit_max_lum->text();
+    (*_new_param)[CurParamIndex::MAX_CLL] = ui->lineEdit_max_cll->text();
+    (*_new_param)[CurParamIndex::MAX_FALL] = ui->lineEdit_max_fall->text();
+    (*_new_param)[CurParamIndex::MASTER_DISPLAY] = QString::number(ui->comboBox_master_disp->currentIndex());
+    (*_new_param)[CurParamIndex::CHROMA_COORD] = ui->lineEdit_chroma_coord->text();
+    (*_new_param)[CurParamIndex::WHITE_COORD] = ui->lineEdit_white_coord->text();
+    (*_new_param)[CurParamIndex::AUDIO_CODEC] = QString::number(ui->comboBox_audio_codec->currentIndex());
+    (*_new_param)[CurParamIndex::AUDIO_BITRATE] = QString::number(ui->comboBox_audio_bitrate->currentIndex());
+    (*_new_param)[CurParamIndex::ASAMPLE_RATE] = QString::number(ui->comboBox_audio_sampling->currentIndex());
+    (*_new_param)[CurParamIndex::ACHANNELS] = QString::number(ui->comboBox_audio_channels->currentIndex());
+    (*_new_param)[CurParamIndex::REP_PRIM] = QString::number(ui->checkBox_primaries->checkState());
+    (*_new_param)[CurParamIndex::REP_MATRIX] = QString::number(ui->checkBox_matrix->checkState());
+    (*_new_param)[CurParamIndex::REP_TRC] = QString::number(ui->checkBox_transfer->checkState());
 
     this->accept();
 }
