+++ app/preset.h
@@ -17,13 +17,13 @@
 #include <QDialog>
 #include <QTimer>
 #include <QListView>
-#include <iostream>
-#include <math.h>
 #include <QMouseEvent>
 #include <QHoverEvent>
 #include <QCloseEvent>
 #include <QResizeEvent>
-
+#include <iostream>
+#include <math.h>
+#include "constants.h"
 
 
 namespace Ui
@@ -102,8 +102,6 @@ private:
 
     Ui::Preset *ui;
 
-    static const int NUMBER_PRESETS = 29;
-
     QVector<QString> *_new_param;
 
     QTimer *timer;
@@ -113,11 +111,6 @@ private:
     float _aspectRatio;
 
     /**************** Geometry **************************/
-
-    enum Resize {
-        LEFT,   LEFT_TOP,   LEFT_BOTTOM,   TOP,
-        RIGHT,  RIGHT_TOP,  RIGHT_BOTTOM,  BOTTOM
-    };
 
     bool        _expandWindowsState,
                 _clickPressedFlag;
@@ -139,33 +132,6 @@ private:
 
     void lockSignals(bool status);
 
-    enum Profile {
-        HIGH,           MAIN,           MAIN10,
-        MAIN12,         PROFILE_0,      PROFILE_1,
-        PROFILE_2,      PROFILE_3,      PROFILE_4,
-        PROFILE_5,      DNXHR_LB,       DNXHR_SQ,
-        DNXHR_HQ,       DNXHR_HQX,      DNXHR_444,
-        PROFILE_AUTO
-    };
-    enum Pixformat {
-        YUV420P12LE,    YUV444P10LE,    YUV422P10LE,
-        YUV420P10LE,    YUV422p,        YUV420P,
-        P010LE,         PIXFORMAT_AUTO
-    };
-    enum Parameters {
-        _OUTPUT_PARAM,   _CODEC,          _MODE,
-        _CONTAINER,      _BQR,            _MAXRATE,
-        _BUFSIZE,        _FRAME_RATE,     _BLENDING,
-        _WIDTH,          _HEIGHT,         _PASS,
-        _PRESET,         _COLOR_RANGE,    _MIN_LUM,
-        _MAX_LUM,        _MAX_CLL,        _MAX_FALL,
-        _MASTER_DISPLAY, _CHROMA_COORD,   _WHITE_COORD,
-        _AUDIO_CODEC,    _AUDIO_BITRATE,  _MINRATE,
-        _LEVEL,          _ASAMPLE_RATE,   _ACHANNELS,
-        _MATRIX,         _PRIMARY,        _TRC,
-        _PRESET_NAME,    _REP_PRIM,       _REP_MATRIX,
-        _REP_TRC
-    };
 };
 
 #endif // PRESET_H
